Migrate element creation to html.js
